+++ Comms_Core/Protos/arty.proto
@@ -29,14 +29,16 @@ message Coords
 
 message ClientReport
 {
-	int32 ClientType = 1; // 0 == Invalid, 1 == Spotter, Client == 2
+	int32 ClientType = 1; // 0 == Invalid, 1 == Spotter, Gunner == 2
 	string SpotterPassword = 2;
 	int32 LastCoordsIdReceived = 3;
+	int32 LastCoordsIdSent = 4; // For use by spotter if they are client.
 }
 
 message ServerReport
 {
 	int32 ActiveCallsignsCount = 1;
 	repeated string ActiveCallsigns = 2;
-	int32 LastCoordsIdSent = 3;
+	int32 LastCoordsIdReceived = 3;
+	int32 LastCoordsIdSent = 4;
 }
