+++ PitBoss/PitBossConstants.cs
@@ -8,7 +8,7 @@ namespace PitBoss
 {
     public static class PitBossConstants
     {
-        public static readonly string VERSION_STRING = "0.3.1";
+        public static readonly string VERSION_STRING = "0.4.0";
         public static int SERVER_PORT = 50082;
         public static int REMOTE_USER_TIMEOUT_MILLISECONDS = 5000;
         //public enum ConnectionStatus
